requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
Django==5.2.4
django-bootstrap5==25.1
pillow==11.3.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
qunicorne==23.0.0